requirements.txt
flask
flask-cors
tensorflow
opencv-python
numpy
Pillow
matplotlib
joblib
keras
streamlit